pyproject.toml
[tool.poetry]
name = "ai_suite"
version = "0.1.0"
description = "demo project"
authors = ["Halyna <halyna.galanzina@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
datasets = "^3.1.0"
pandas = "^2.2.3"
instructor = "^1.6.3"
openai = "^1.53.0"
getenv = "^0.2.0"
matplotlib = "^3.9.2"
pydantic = "^2.10.5"
aiohttp = "^3.11.11"
anthropic = "^0.42.0"
google-generativeai = "^0.8.3"
tabulate = "^0.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
